Separating stories into own files
